dev(qbus): one more step to qbus
