+++ CMakeLists.txt
@@ -25,12 +25,10 @@ endif (Boost_FOUND)
 
 set(CMAKE_CXX_FLAGS "-Wall -Werror")
 add_subdirectory(qbus)
-add_subdirectory(qbus_way)
 if (QBUS_EXAMPLES_ENABLED)
     add_subdirectory(examples)
 endif (QBUS_EXAMPLES_ENABLED)
 if (Boost_FOUND AND QBUS_TEST_ENABLED)
     enable_testing()
     add_subdirectory(tests)
-    add_subdirectory(qbus_tests)
 endif (Boost_FOUND AND QBUS_TEST_ENABLED)
