+++ tests/connector_test.cpp
@@ -26,13 +26,13 @@ BOOST_AUTO_TEST_CASE(simple_test)
     pconnector_type pconnector2 = connector::make<in_connector_type>("test");
     BOOST_REQUIRE(pconnector1);
     BOOST_REQUIRE(pconnector2);
-    BOOST_REQUIRE(pconnector1->create(512, 32));
+    BOOST_REQUIRE(pconnector1->create(0, 32 * 512));
     BOOST_REQUIRE(pconnector2->open());
     BOOST_REQUIRE(!pconnector1->get());
     BOOST_REQUIRE(!pconnector1->pop());
     BOOST_REQUIRE(!pconnector2->get());
     buffer_t buffer = make_buffer(512);
-    BOOST_REQUIRE(pconnector1->add(&buffer[0], buffer.size()));
+    BOOST_REQUIRE(pconnector1->add(0, &buffer[0], buffer.size()));
     pmessage = pconnector2->get();
     BOOST_REQUIRE(pmessage);
     BOOST_REQUIRE_EQUAL(pmessage->data_size(), buffer.size());
